+++ tests/testthat/test-ots_product_code.R
@@ -4,13 +4,13 @@ test_that("ots_product_code works properly for a partial product string matching
   test_product <- ots_product_code(productname = "fruit")
 
   expect_is(test_product, "tbl")
-  expect_output(str(test_product), "26 obs")
-  expect_output(str(test_product), "5 variables")
+  expect_equal(nrow(test_product),26)
+  expect_equal(ncol(test_product),3)
 })
 
 test_that("ots_product_code returns 0 rows for a non-existing product match", {
   d <- ots_product_code(productname = "adamantium")
-  expect_output(str(d), "0 obs")
+  expect_equal(nrow(d),0)
 })
 
 test_that("ots_product_code returns an error when no product is specified", {
@@ -21,13 +21,13 @@ test_that("ots_product_code works properly for a partial group string matching",
   test_group <- ots_product_code(productgroup = "vegetable")
   
   expect_is(test_group, "tbl")
-  expect_output(str(test_group), "58 obs")
-  expect_output(str(test_group), "5 variables")
+  expect_equal(ncol(test_group),3)
+  expect_equal(nrow(test_group),6)
 })
 
 test_that("ots_product_code return 0 rows for a non-existing group match", {
   d <- ots_product_code(productgroup = "headphones and speakers")
-  expect_output(str(d), "0 obs")
+  expect_equal(nrow(d),0)
 })
 
 test_that("ots_product_code returns an error when no group is specified", {
@@ -38,8 +38,8 @@ test_that("ots_product_code works ok for both specified product and group", {
   test_both <- ots_product_code(productname = "potato", productgroup = "vegetable")
   
   expect_is(test_both, "tbl")
-  expect_output(str(test_both), "2 obs")
-  expect_output(str(test_both), "6 variables")
+  expect_equal(ncol(test_both),6)
+  expect_equal(nrow(test_both),2)
 })
 
 test_that("ots_product_code fails with NULL product/group", {
@@ -57,5 +57,5 @@ test_that("ots_product_code returns error or no results for strange inputs", {
   # this shall return an empty data.frame
   d <- ots_product_code(productname = "kriptonite", productgroup = "adamantium")
   expect_is(d, "tbl")
-  expect_output(str(d), "0 obs")
+  expect_equal(nrow(d),0)
 })