+++ NEWS.md
@@ -1,3 +1,13 @@
+# Version 0.4.0
+
+Updates
+
+* Includes `yrpc-ga`, `yrpc-sa`, `yrc-ga` and `yr-sa` tables reflecting API updates
+* Simplifies end-user functions a bit (i.e. removes `include_groups` option)
+* Optimizes the code a bit, specially at the joins with tables in the package
+* Fixes codes duplication when both product and group/community match for a search
+* Includes both official and shortened section names
+
 # Version 0.3.1
 
 Updates